NODE-2246: Fix blockchain-updates work with importer (#3282)
