+++ node/src/main/scala/com/wavesplatform/Importer.scala
@@ -99,6 +99,8 @@ object Importer extends ScorexLogging {
     settings.copy(dbSettings = settings.dbSettings.copy(useBloomFilter = true))
   }
 
+  private[this] var triggers = Seq.empty[BlockchainUpdateTriggers]
+
   def initExtensions(
       wavesSettings: WavesSettings,
       blockchainUpdater: BlockchainUpdaterImpl,
@@ -153,6 +155,7 @@ object Importer extends ScorexLogging {
       extensions.flatMap { ext =>
         Try(ext.start()) match {
           case Success(_) =>
+            triggers ++= Some(ext).collect { case t: BlockchainUpdateTriggers => t }
             Some(ext)
           case Failure(e) =>
             log.warn(s"Can't initialize extension $ext", e)
@@ -247,20 +250,19 @@ object Importer extends ScorexLogging {
     val scheduler = Schedulers.singleThread("appender")
     val time      = new NTP(settings.ntpServer)
 
-    val actorSystem              = ActorSystem("wavesplatform-import")
-    val db                       = openDB(settings.dbSettings.directory)
+    val actorSystem = ActorSystem("wavesplatform-import")
+    val db          = openDB(settings.dbSettings.directory)
     val (blockchainUpdater, levelDb) =
-      StorageFactory(settings, db, time, Observer.empty, BlockchainUpdateTriggers.noop)
+      StorageFactory(settings, db, time, Observer.empty, BlockchainUpdateTriggers.combined(triggers))
     val utxPool     = new UtxPoolImpl(time, blockchainUpdater, PublishSubject(), settings.utxSettings)
     val pos         = PoSSelector(blockchainUpdater, settings.synchronizationSettings)
     val extAppender = BlockAppender(blockchainUpdater, time, utxPool, pos, scheduler, importOptions.verify) _
 
-    checkGenesis(settings, blockchainUpdater)
     val extensions = initExtensions(settings, blockchainUpdater, scheduler, time, utxPool, db, actorSystem)
+    checkGenesis(settings, blockchainUpdater)
 
     sys.addShutdownHook {
       quit = true
-      Await.ready(Future.sequence(extensions.map(_.shutdown())), settings.extensionsShutdownTimeout)
       Await.result(actorSystem.terminate(), 10.second)
       lock.synchronized {
         if (blockchainUpdater.isFeatureActivated(BlockchainFeatures.NG)) {
@@ -283,7 +285,14 @@ object Importer extends ScorexLogging {
           )
           blockchainUpdater.processBlock(pseudoBlock, ByteStr.empty, verify = false)
         }
+
+        // Terminate appender
         scheduler.shutdown()
+        scheduler.awaitTermination(10 seconds)
+
+        // Terminate extensions
+        Await.ready(Future.sequence(extensions.map(_.shutdown())), settings.extensionsShutdownTimeout)
+
         blockchainUpdater.shutdown()
         levelDb.close()
         db.close()
